+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.5.1
 
 - `Flyout` fixed compatibility with IE11
+- `Headline` styles update
+- Updated label in `Modal`
 
 ## 0.5.0